+++ tests/testthat/test-mark.R
@@ -180,9 +180,9 @@ describe("summary.bench_mark", {
 
 describe("unnest.bench_mark", {
   it("does not contain result or memory columns", {
-    skip_if_not_installed("tidyr")
+    skip_if_not_installed("tidyr", minimum_version = "0.8.99")
     bnch <- mark(1+1, 2+0)
-    res <- tidyr::unnest(bnch)
+    res <- tidyr::unnest(bnch, c(time, gc))
 
     gc_cols <- colnames(bnch$gc[[1]])